Emulate RCS $Id$ with git describe.
A killer feature of RCS was the $Id$ tag, which both enabled you to
integrate the file revision version/identity in a file and also came
with the nifty ident utility that shows you those tags in files, even
binaries.

Emulate $Id$ with a define created by git describe, simply to be able
to still use ident when finding a binary module and wondering which source
version it corresponds to.